% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{head.pash}
\alias{head.pash}
\title{Head function for pash object}
\usage{
head.pash(object, n = 6L)
}
\arguments{
\item{object}{Pash object.}

\item{n}{A single integer. If positive, size for the resulting object: number of rows for a life-table. If negative, all but the n last/first number of elements of x.}
}
\description{
\emph{bold{Generic function}}
}

