% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pash_correction.r
\name{machine.round}
\alias{machine.round}
\title{Rounding a number to the machine precission to avoid numerical mistakes during comparison}
\usage{
machine.round(x)
}
\description{
Rounding a number to the machine precission to avoid numerical mistakes during comparison
}
\keyword{internal}

