% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.aggregate}
\alias{pclm.aggregate}
\title{Calculate Raw and Aggregated Life-tables}
\usage{
pclm.aggregate(fit, out.step = NULL, count.type = c("DX", "LX"),
  exposures.used = FALSE)
}
\arguments{
\item{fit}{Object obtained by \code{\link{pclm.opt}} function.}

\item{out.step}{Output interval length.}

\item{count.type}{Type of the data, deaths(\code{"DX"})(default) or
exposures(\code{"LX"}.)}

\item{exposures.used}{Logical indicating if exposures were used to create
composition matrix.}
}
\value{
List with components:

\item{\code{grouped}}{Life-table based on aggregated PCLM fit defined
  by \code{out.step}.}

\item{\code{raw}}{Life-table based on original (raw) PCLM fit.}

\item{\code{fit}}{PCLM fit used to construct life-tables.}

\item{\code{warn.list}}{List with warnings.}
}
\description{
Calculate raw and aggregated life-tables From the object returned by the
\code{pclm.opt()} Function.
}
\seealso{
\code{\link{pclm.general}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}>
  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
