% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.compmat}
\alias{pclm.compmat}
\title{Create composition matrix object}
\usage{
pclm.compmat(x, y, exposures = NULL, control = list())
}
\arguments{
\item{x}{Vector with start of the interval for age/time classes. x * x.div must be an integer.
The appropriate correction for fractional intervals based on the interval multiple (\code{\link{pclm.interval.multiple}}) is performed in \code{\link{pclm.general}}.}

\item{y}{Vector with counts, e.g. \code{ndx}. It must have the same length as \code{x}.}

\item{exposures}{Vector with exposures used to calculate smoothed mortality rates (see reference [1] and \code{\link{pclm.general}}).}

\item{control}{List with additional parameters. See \code{\link{pclm.control}}.}
}
\value{
List with components:

\item{\code{C}}{ Composition matrix.}

\item{\code{X}}{ B-spline base, P-spline base, or identity matrix.}

\item{\code{x}}{ Corrected age/time vector.}

\item{\code{y}}{ Corrected vector with counts.}

\item{\code{open.int.len}}{ Length of the open interval in age classes.}

\item{\code{exposures}}{ Vector with exposures if it was used to construct the composition matrix.}

\item{\code{control}}{Used control parameters, see \code{\link{pclm.control}}.}

\item{\code{warn.list}}{ List with warnings.}
}
\description{
Construct the composition matrix object for automatically recalibrated age classes. \cr\cr \emph{\bold{The internal function}}.
}
\details{
The details of matrix construction can be found in reference [1].
if \code{bs.use == TRUE} then P- or B- splines are used instead of identity matrix (see \code{\link{pclm.control}}).\cr\cr
The dimension of constructed composition matrix can be determined before its computation.
The shorter dimension equals to the length of data vector + 1, whereas the longer dimension is
determined by the function \code{\link{pclm.nclasses}} and for \code{zero.class.end == NULL} equals:\cr\cr
\code{(x.div * (max(x) - min(x)) * m) * (1 + zero.class.frac) + 1}\cr\cr
or\cr\cr
\code{x.div * (zero.class.end - min(x)) * m + 1}\cr\cr
otherwise,
where \code{m} is an interval multiple calculated by \code{\link{pclm.interval.multiple}}.
See also \code{\link{pclm.nclasses}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\references{
\enumerate{
\item{Rizzi S, Gampe J, Eilers PHC. Efficient estimation of smooth distributions from coarsely grouped data. Am J Epidemiol. 2015;182:138?47.}
\item{Camarda, C. G. (2012). MortalitySmooth: An R Package for Smoothing Poisson Counts with P-Splines. Journal of Statistical Software. 50, 1-24.}
\item{Hastie, T. J. (1992) Generalized additive models. Chapter 7 of Statistical Models in S eds J. M. Chambers and T. J. Hastie, Wadsworth & Brooks/Cole.}
}
}
\seealso{
\code{\link{pclm.general}}, \code{\link{pclm.control}}, \code{\link{pclm.interval.multiple}}, and \code{\link{pclm.nclasses}}.
}
\keyword{internal}

