% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.control}
\alias{pclm.control}
\title{Auxiliary for controlling PCLM fitting}
\usage{
pclm.control(x.div = 1L, x.auto.trans = TRUE, x.max.ext = 25L,
  zero.class.add = TRUE, zero.class.end = NULL, zero.class.frac = 0.2,
  bs.use = "auto", bs.method = c("MortalitySmooth", "bs"),
  bs.df = c("maxprec", "thumb"), bs.df.max = 200L, bs.deg = 3L,
  opt.method = c("BIC", "AIC"), opt.tol = .Machine$double.eps^0.5,
  pclm.deg = 2L, pclm.max.iter = 100L,
  pclm.lsfit.tol = .Machine$double.eps^0.5,
  pclm.tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x.div}{Number of sub-classes within PCLM tim/age class (default is 1).
Low value of the parameter makes the PCLM computation faster. It is however recommended to set
it to higher value (e.g. 10) for better \code{nax} estimates.}

\item{x.auto.trans}{Logical indicating if automatically multiple age intervals to remove fractions.
\code{TRUE} is the recommended value. See also examples in \code{\link{pclm.fit}}.}

\item{x.max.ext}{Integer defining maximal multiple of an age interval. See also \code{\link{pclm.interval.multiple}}.}

\item{zero.class.add}{Logical indicating if additional zero count class (open interval)
should be added after last age class. \code{TRUE} is the recommended value. See \code{\link{pclm.nclasses}} and \code{\link{pclm.compmat}}.}

\item{zero.class.end}{Positive indicating the end of zero count class = anticipated end of
last (open) interval. If set to \code{NULL} and \code{zero.class.add == TRUE} then it is calculated
automatically based on \code{zero.class.frac}. See \code{\link{pclm.nclasses}} and \code{\link{pclm.compmat}}.}

\item{zero.class.frac}{Fraction of total range of \code{x} (age/time vector) added as the last zero-count interval
when \code{zero.class.end == NULL}. Used in \code{\link{pclm.compmat}}. Increase this value if the right tail
of the PCLM fit is badly fitted (use \code{\link{plot.pclm}} to diagnose).}

\item{bs.use}{Logical indicating if use B- or P-spline basis to speed-up computations.
Possible values: \code{"auto"}, \code{TRUE}, and \code{FALSE}. Used by \code{\link{pclm.compmat}} function.}

\item{bs.method}{Basis for B- or P-spline used by \code{\link{pclm.compmat}} function.
Possible values:
\itemize{
\item{\code{"MortalitySmooth"}}{ - gives "P-splines" basis based on \code{\link{MortSmooth_bbase}}
\code{\{\link{MortalitySmooth}\}} (recommended)}
\item{\code{"bs"}}{ - gives basic B-splines basis based on \code{\link{bs}} \code{\{\link{splines}\}}.}
}}

\item{bs.df}{B- or P- spline degree of freedom (df, number of inner knots)
or a way to its calculation used in \code{\link{pclm.compmat}} function.
The value is automatically limited by the \code{bs.df.max}.
It can take corresponding values:
\itemize{
\item{\code{"maxprec"}}{ - df equal to the number of ungrouped (raw) age classes (recommended option).}
\item{\code{"thumb"}}{ - 'rule of thumb': one knot for the B-spline basis each 4-5 observations.}
\item{\code{integer}}{ - df given explicitly.}
}}

\item{bs.df.max}{Maximal number of knots (df) for B- or P-spline basis.
Defaut value is 200, but can be decreased if computations are slow.
Used in \code{\link{pclm.compmat}}.}

\item{bs.deg}{Degree of the piecewise polynomial for B- or P-spline basis.
Default and recommended value is 3. Used in \code{\link{pclm.compmat}}.}

\item{opt.method}{Selection criterion for \code{lambda} (smooth parameter) in
\code{\link{pclm.opt}} function. Possible values are \code{"AIC"} and \code{"BIC"} (recommended).}

\item{opt.tol}{Tolerance for \code{\link{pclm.opt}} function that estimates smooth
parameter \code{lambda}.}

\item{pclm.deg}{Order of differences of the components of \code{b} (PCLM coeficients, beta in the reference [1]).
Default value is 2, some other values may cause algorithm to not
work properly. Used by \code{\link{pclm.core}} function.}

\item{pclm.max.iter}{Maximal number of iterations in \code{\link{pclm.core}}
function. Default is 100, but increase it when got a warning.}

\item{pclm.lsfit.tol}{Tolerance for \code{\link{lsfit}} function used
in \code{\link{pclm.core}} function.}

\item{pclm.tol}{Tolerance for \code{\link{pclm.core}} function.}
}
\value{
List with control parameters.
}
\description{
Auxiliary function for controlling PCLM fitting. Use this function to set control
parameters of the \code{\link{pclm.fit}} and other related functions.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\references{
\enumerate{
\item{Rizzi S, Gampe J, Eilers PHC. Efficient estimation of smooth distributions from
coarsely grouped data. Am J Epidemiol. 2015;182:138?47.}
}
}
\seealso{
\code{\link{pclm.fit}}, \code{\link{pclm.general}}, \code{\link{pclm.core}},
\code{\link{pclm.opt}}, \code{\link{pclm.aggregate}}, \code{\link{pclm.compmat}},
\code{\link{pclm.interval.multiple}}, \code{\link{pclm.nclasses}},
\code{\link{plot.pclm}}, and \code{\link{summary.pclm}}.
}

