% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.interval.multiple}
\alias{pclm.interval.multiple}
\title{Multiple for the original age/time interval length}
\usage{
pclm.interval.multiple(x, control = list())
}
\arguments{
\item{x}{Vector with start of the interval for age/time classes.}

\item{control}{List with additional parameters. See \code{\link{pclm.control}}.}
}
\description{
Calculates minimal multiple of the orginal age/time interval length to remove fractions in \code{x}.
The function (together with \code{x} and some control parameters) is used to calculate the
internal (raw, nonaggregated) interval length during PCLM computations.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\seealso{
\code{\link{pclm.general}}, \code{\link{pclm.control}}, and \code{\link{pclm.nclasses}}.
}

