% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.opt}
\alias{pclm.opt}
\title{Optimize the smooth parameter \code{lambda} for PCLM method}
\usage{
pclm.opt(CompositionMatrix, control = list())
}
\arguments{
\item{CompositionMatrix}{Output of \code{\link{pclm.compmat}}.}

\item{control}{List with additional parameters. See \code{\link{pclm.control}}.}
}
\value{
List with components:

\item{\code{X}}{Ungrouped age/time classes.}

\item{\code{Y}}{Ungrouped counts.}

\item{\code{lambda}}{Optimal smooth parameter.}

\item{\code{fit}}{Output of the \code{\link{pclm.core}} derived for the optimal \code{lambda}.}

\item{\code{CompositionMatrix}}{Used composition matrix. See also \code{\link{pclm.compmat}}.}

\item{\code{warn.list}}{List with warnings.}
}
\description{
\emph{\bold{The internal function}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\references{
\enumerate{
\item{Rizzi S, Gampe J, Eilers PHC. Efficient estimation of smooth distributions from coarsely grouped data. Am J Epidemiol. 2015;182:138?47.}
}
}
\keyword{internal}

