% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{summary.pclm}
\alias{summary.pclm}
\title{Summary of the Fitted PCLM Object}
\usage{
\method{summary}{pclm}(object)
}
\arguments{
\item{object}{Fitted PCLM object.}
}
\description{
Summary of the Fitted PCLM Object
}
\seealso{
\code{\link{pclm.fit}} \code{\link{plot.pclm}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}>
  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
