% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{pclm.aggregate}
\alias{pclm.aggregate}
\title{Calculate raw and aggregated life table from the object returned by \code{pclm.opt()} function.}
\usage{
pclm.aggregate(fit, out.step = NULL, count.type = c("DX", "LX"),
  exposures.used = FALSE)
}
\arguments{
\item{fit}{Object obtained by \code{\link{pclm.opt}} function.}

\item{out.step}{Output interval length.}

\item{count.type}{Type of the data, deaths(\code{"DX"})(default) or exposures(\code{"LX"}.)}

\item{exposures.used}{Logical indicating if exposures were used to create composition matrix.}
}
\value{
List with components:

\item{\code{grouped}}{Life-table based on aggregated PCLM fit defined by \code{out.step}.}

\item{\code{raw}}{Life-table based on original (raw) PCLM fit.}

\item{\code{fit}}{PCLM fit used to construct life-tables.}

\item{\code{warn.list}}{List with warnings.}
}
\description{
\emph{\bold{The internal function}}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\seealso{
\code{\link{pclm.fit}}
}
\keyword{internal}

