% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{print.pclm}
\alias{print.pclm}
\title{Print function for PCLM object}
\usage{
print.pclm(object, type = c("lt", "aggregated", "nonaggregated"), ...)
}
\arguments{
\item{object}{PCLM object.}

\item{type}{which life-table  should be returned. One of \code{"lt"}, \code{"aggregated"} or \code{"nonaggregated"}.}

\item{...}{other parameters passed to \code{\link{print}}.}
}
\description{
\emph{bold{Generic function}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

