% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm_pash.r
\name{summary.pclm}
\alias{summary.pclm}
\title{Summary of the fitted PCLM object}
\usage{
summary.pclm(object)
}
\arguments{
\item{object}{Fitted PCLM object.}
}
\description{
\emph{bold{Generic function}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\seealso{
\code{\link{pclm.fit}} \code{\link{plot.pclm}}
}
\keyword{internal}

